Extract ReadReservationRow helper method
I usually don't extract helper methods when only two duplicates exist
(re: The Rule of Three), but there are other motivations for creating
helper methods:

The most common alternative motivation is to give a subprocedure a
descriptive name.

Another, releated motivation is to reduce the size of the calling
method. That's the motivation in play here.