Introduce Restaurant class
It's not yet used, but the intent is to let it replace the
Options.Restaurant that's infected the test code too much. Also, I think
this will simplify interaction with the restaurant database in the
Controllers, because they'll be able to (in the future) retrieve a
restaurant object in one go, instead of first having to query for the
name, and then for the maitre d'.