Extract TryUpdate helper method
The Put method had grown too large. Even this commit, splitting the code
into two approximately equally sized code blocks, leaves both of them
larger than 24 lines of code.

This refactoring is opportunistic rather than following a particular
principle. You could say that it's another example of the art of
software engineering. I was first and foremost aiming for pulling out
everything that goes on between the creation and the completion of the
transaction scope, so that it would be clear to a code reader that a
scope was active.

This does leave the Put method at a heterogeneous level of abstraction,
because what goes on before the scope kicks in is more fine-grained than
what happens inside the scope.

Since the Put method is still too big, though, it'd be obvious to look
for a way to collapse all that preparation code to a single helper
method as well.