Rename `res` to `reservation`
The short variable name `res` could mean 'restaurant' just as much as it
might mean 'reservation', that's obviously an ambiguous name. As Clean
Code's N5 heuristic suggests, short names are fine for short scopes, but
longer names should be used for longer scopes.

I'm not in the mood to do source code archeology on the variable name
`res` in this context, but I'd guess that I named the variable at a time
when the scope was smaller. It might have been a good name once, but now
it isn't.