Rename `res` to `reservation`
The short variable name `res` could mean 'restaurant' just as much as it
might mean 'reservation', that's obviously an ambiguous name. As Clean
Code's N5 heuristic suggests, short names are fine for short scopes, but
longer names should be used for longer scopes.

This didn't improve the size of the method, though.