Revert "Extrace TryUpdate helper method"
This reverts commit 4c6fdaa134b637535164b87a709b8b3a72a872e3.

The reason I'm reverting the commit is that while it illustrates an
opportunistic refactoring, it also shows how this can sometimes lead to
a dead end. This is also the reason I'm leaving the original commit in
place, so that you can compare and contrast.

The problem is that if you extract a helper method like TryUpdate, it's
difficult to move further. You can attempt to implement a Chain of
Responsibility that returns a ActionResult if it can. Because of the
presence of the transaction scope, however, you'll need to be able to
complete the scope if and only if no ActionResult has been produced so
far. In other words, you need to be able to detect whether the next
handler in the Chain of Responsibility succeeds or fails. This may
either lead towards various ad-hoc half-implementations of TryDo or
Tester-Doers, but neither seems particularly promising.

I started down the Chain of Responsibility path before I realised this;
I haven't tried the other options.

Left is, it seems, the Either sum type, which I will now attempt to use
instead.