Add Either
This is a Visitor-based implementation of the Either monad. A less
verbose implementation can be created using Church-encoding
(https://blog.ploeh.dk/2018/06/11/church-encoded-either). I've done this
so many times that I don't feel the need to add tests, but you can, if
you will. It's easy enough to unit test (I've done this many times as
well).