Extract TryUpdate method
This looks like the same sort of refactoring as before, but I know from
years of experience with Either in both C#, F#, and Haskell that this is
a universal abstraction that composes (and decomposes).

It's not native to C#, so syntactically, it may require a bit of getting
used to. Also, Unit is a type isomorphic to void, but a proper type
that can be used as a generic type argument.