Extract WillAcceptUpdate helper method
The whole Put flow is now getting so drawn out that it might be
worthwhile considering refactoring to introduce a separate object that
handles that action. This might reduce the need to pass the same
arguments along between method calls...

First, however, I'll see how much more refactoring work remains on the
production code. It might not be worthwhile yet.