Extract AssertOverbookResponse helper assertion
I admit that the main motivation for this refactoring is to reduce the
line count of the OverbookAttempt test method. This is, perhaps,
dubious.

You could argue, though, that the expected response to an overbook
attempt is controversial enough to warrant a bit of encapsulation. The
choice of returning 500 Internal Server error when the expected
condition of being sold out occurs rubs many people the wrong way. I
think I may briefly have discussed this somewhere, but in this context,
the point is that this API design decision might change (watch out for
breaking changes, though), so encapsulating the assertion is a helper
method may make a bit of sense...