Replace underlying class field with auto-property
This is the last change I plan to make on Table. The external API is
exactly the same, but the internal implementation went from 187 lines of
code to 87 lines of code.

This isn't without drawbacks, but I think that the advantages (less
code) outweigh the disadvantages (less pattern-based implementation).
Perhaps the disadvantage require more explanation:

The previous example was based on the Visitor pattern. While this
pattern isn't the most widely known pattern, for those that know it, the
implementation shouldn't be much of a surprise. With the new
implementation, programmers who know the Visitor design pattern may
expect to encounter an implementation based on it, only to find if/else
branching instead. While this can be confusing for that subset of
programmers, I hope that overall, the change is to the better.