Add restarant ID to Delete method
While this parameter currently isn't required, that's actually a kind of
reverse or negative leaky abstraction. The API for the
IReservationsRepository interface is inconsistent, then, because some
methods take a restaurant ID and other don't. This commit rectifies one
of these inconsistencies.

Apart from the above abstract or theoretical reasons, the lack of a
restaurant ID also prevents some nice refactorings, such as making the
post office a Decorator of the repository.