Introduce TimeSlot class
It turns out that I've over-complicated the timeslot API by making
Occurrence a generic container. Despite doing that, I never used it for
anything than a container of tables. Using a more concrete TimeSlot
class may simplify the code.

The new class is not yet in use, but I plan to move things over
piecemeal in the next commits, according to the Strangler pattern.