Move static mutation out of lambda expression
It turns out that I'd had a bug in the code for four days without
noticing it. No unit tests broke; the build didn't break. I only
discovered the issue because I tried to use the feature in a real
deployment. Clearly, I hadn't been testing that feature in the live
environment for the last four days. This is, after all, example code,
and no users actually use the deployed system. In a real setting, this
defect would probably not had gone unnoticed quite as long.

The problem was that I couldn't GET the schedule from the live system.
After some incredulous but-all-tests-pass troubleshooting, I finally
decided to try to reproduce the problem on my laptop. While all tests
were passing, I couldn't GET the schedule when I stood up the system on
my development machine. In both cases, the symptom was the same: a 403
Forbidden response.

I wasted some time trying to 'decompile' the token I'd been using, but
it looked fine. Putting breakpoints in the source code didn't help,
which suggested that the culprit somehow was the [Authorize] attribute
on the Controller. Removing that attribute made the problem disappear,
so that seemed to corroborate that hypothesis.

After some more time getting nowhere, I discovered that while all tests
were passing, if I ran ONLY the test method
GetRestaurantScheduleWhileAuthorized, one of the three test cases would
fail. After still more wasted time, I decided to search through the Git
history to see if I could identify the commit that introduced the
defect. Since the build was passing, I couldn't use automated git
bisect, but doing it by hand wasn't too hard either. After eight
checkouts, I'd found the commit that introduced the problem.

It's the commit with the title "Extract CreateTokenValidationParameters
method" (I'm not going to refer to commit IDs, because I'm probably end
up rebasing this repository later).

One of many benefits of small commits is that once you find the
offending commit, it's easy to spot the problem. That particular commit
didn't involve more than a dozen lines of code. It's a good thing I'm
not in the habit of squashing commits.

The problem, it turns out, is mutation of shared mutable state. Who
could have guessed?! I thought that the line:

JwtSecurityTokenHandler.DefaultMapInboundClaims = false;

fit best within the CreateTokenValidationParameters method, because it
was directly related to that, but what I hadn't realised was that
CreateTokenValidationParameters is called from within a lambda
expression. In other words, it doesn't run immediately.

This seems to explain that when you run it together whith a larger suite
of tests, other tests will already have exercised that code path and
caused the shared state to change. If, on the other hand, you run the
code in isolation, it doesn't run.

I don't see an easy way to reproduce this defect as an automated test.
It's a great candidate for an automated system test (against a deployed
system), but this code based doesn't include such tests.