Remove redundant code analysis suppression
I'm not sure, and I haven't investigated, but Visual Studio came out and
told me in a friendly way that this suppression wasn't necessary. I
wonder if this is because of an updated analysis rule.

I seem to recall that there was an open issue exactly about the type of
warning suppressed here: that GUIDs aren't culture-specific, so that the
warning aught not to have been issued. I guess that the team that
maintains the code analysis engine added a special case. When I wrote
the code, that special case wasn't in the engine, but now it is. That's
my guess, anyway.