Make test relative to system clock
The test in question was going to fail after March 18, 2022, since one
of the hard-coded dates in the [InlineData] attributes would then be in
the past and be rejected by the API (which rejects reservations in the
past).

This creates other problems, since the size of the test now grew. One
thing at a time, though.