Make test relative to current time
This test would fail after 2022-04-01 because the actual system
(encapsulated by `api`) runs on the system clock. Now it always
runs 437 days in the future.

That number is fairly arbitrary. It's just the number of days that
2022-04-01 is from today's day, as I'm committing this. Had I done it
tomorrow, it'd been 436, and so forth. I feel that this makes a little
sense, though, as I trust that the test was trustworthy today, before
the refactoring. So today, at least, the refactored test is equivalent
to the original test. Tomorrow, they diverge.