Make test relative to current time
This test would fail after 2023-06-24 because the actual system
(encapsulated by `api`) runs on the system clock. Now it always
runs 884 and 902 days in the future.

Those numbers are fairly arbitrary. It's just the number of days that
the hard-coded test cases was from today's day, as I'm committing this.
Had I done it tomorrow, it'd been other numbers.