Make test relative to current time
This test would fail after 2023-04-10 because the actual system
(encapsulated by `api`) runs on the system clock. Now it always
runs 809 days in the future.

Those numbers are fairly arbitrary. It's just the number of days that
the hard-coded test case was from today's day, as I'm committing this.
Had I done it tomorrow, it'd been another number.