Make test cases relative to current date
For detailed explanation of how the test case values for `days` were
picked, see the previous commit messages.

This change unfortunately exceeds the 80x24 box threshold for method
size. This is partly an artefact of the parameter list layout: 7 method
parameters take up 7 lines of code with this layout.

The underlying problem is imposed primitive obsession, as explained by
https://blog.ploeh.dk/2021/01/18/parametrised-test-primitive-obsession-code-smell

One fix would be to refactor the test to use [ClassData] as a source for
test data, instead of [InlineData], as also explained in the
above-mentioned article. I'm currently debating with myself whether
that'd be overkill or not...